Make root support better, make ADB optional.

+++ index.js
@@ -9,15 +9,16 @@ const inquirer = require('inquirer');
 const argParser = opteric(process.argv.join(' '));
 const actualExec = util.promisify(exec);
 
+let adbExists = true;
 const jarNames = {
   cli: './revanced/',
   patchesJar: './revanced/',
   integrations: './revanced/',
-  deviceId: null,
+  deviceId: '',
   microG: './revanced/'
 };
 
-async function overWriteJarNames (link) {
+async function overWriteJarNames(link) {
   const fileName = link.split('/').pop();
   // i have to use ifs for this sorry
   if (fileName.includes('revanced-cli')) jarNames.cli += fileName;
@@ -28,7 +29,7 @@ async function overWriteJarNames (link) {
   if (fileName.startsWith('microg')) jarNames.microG += fileName;
 }
 
-async function getDownloadLink (json) {
+async function getDownloadLink(json) {
   const apiRequest = await fetchURL(
     `https://api.github.com/repos/${json.owner}/${json.repo}/releases/latest`
   );
@@ -36,7 +37,7 @@ async function getDownloadLink (json) {
   return jsonResponse.assets;
 }
 
-async function getPage (pageUrl) {
+async function getPage(pageUrl) {
   const pageRequest = await fetchURL(pageUrl, {
     headers: {
       'user-agent':
@@ -46,12 +47,13 @@ async function getPage (pageUrl) {
   return await pageRequest.text();
 }
 
-async function downloadYTApk () {
+async function downloadYTApk(ytVersion) {
   const versionsList = await getPage(
     'https://www.apkmirror.com/apk/google-inc/youtube'
   );
   const $ = load(versionsList);
-  let apkVersionText = $(
+  let apkVersionText = ytVersion || 
+  $(
     'h5[class="appRowTitle wrapText marginZero block-on-mobile"]'
   ).get()[0].attribs.title;
 
@@ -100,7 +102,7 @@ async function downloadYTApk () {
   return console.log('Download complete!');
 }
 
-async function downloadFile (assets) {
+async function downloadFile(assets) {
   for (const asset of assets) {
     const dir = fs.readdirSync('./revanced/');
     overWriteJarNames(asset.browser_download_url);
@@ -124,20 +126,20 @@ async function downloadFile (assets) {
   }
 }
 
-async function downloadFiles (repos) {
+async function downloadFiles(repos) {
   for (const repo of repos) {
     const downloadLink = await getDownloadLink(repo);
     await downloadFile(downloadLink);
   }
 }
 
-async function getADBDeviceID () {
+async function getADBDeviceID() {
   let deviceId;
   const { stdout } = await actualExec('adb devices');
   const match = stdout.match(/^(\w+)\s+device$/m);
   if (match === null) {
     console.log('No device found! Fallback to only build.');
-    return (jarNames.deviceId = '');
+    return;
   }
 
   const [deviceIdN] = match;
@@ -145,7 +147,7 @@ async function getADBDeviceID () {
   return deviceId;
 }
 
-async function checkForJavaADB () {
+async function checkForJavaADB() {
   try {
     const javaCheck = await actualExec('java -version');
     const javaVer = Array.from(javaCheck.stderr.matchAll(/version\s([^:]+)/g))
@@ -157,7 +159,6 @@ async function checkForJavaADB () {
       );
       return process.exit();
     }
-
     await actualExec('adb');
   } catch (e) {
     if (e.stderr.includes('java')) {
@@ -167,10 +168,18 @@ async function checkForJavaADB () {
     }
     if (e.stderr.includes('adb')) {
       console.log(
-        "You don't have ADB installed.\nPlease get it from here: https://developer.android.com/studio/releases/platform-tools\nIf you are confused on how you're gonna install it, watch some tutorials on YouTube!"
+        "You don't have ADB installed.\nPlease get it from here: https://developer.android.com/studio/releases/platform-tools\nIf you are gonna use rooted ReVanced, this is needed."
       );
+      adbExists = false;
     }
   }
+}
+
+async function getYTVersion() {
+  const { stdout, stderr } = await actualExec('adb shell dumpsys package com.google.android.youtube');
+  const dumpSysOut = stdout || stderr;
+  return dumpSysOut.match(/versionName=([^=]+)/)[1].replace('\r\n    splits', '');
+
 }
 
 (async () => {
@@ -201,6 +210,8 @@ async function checkForJavaADB () {
     }
 
     case 'patch': {
+      let excludedPatches = '';
+      let ytVersion;
       if (!fs.existsSync('./revanced')) {
         fs.mkdirSync('./revanced');
       }
@@ -227,21 +238,24 @@ async function checkForJavaADB () {
       ];
       await downloadFiles(filesToDownload);
 
-      if (!argParser.flags.includes('manual-apk')) {
-        await downloadYTApk();
-      }
-
-      await getADBDeviceID();
-      let excludedPatches = '';
-
       if (argParser.options.exclude) {
         if (argParser.options.exclude.includes('microg-support')) {
+          ytVersion = await getYTVersion();
           excludedPatches += '--mount';
         }
         for (const patch of argParser.options.exclude.split(',')) {
           excludedPatches += ` -e ${patch}`;
         }
       }
+
+      if (!argParser.flags.includes('manual-apk')) {
+        await downloadYTApk(ytVersion);
+      }
+
+      if (adbExists) {
+        await getADBDeviceID();
+      }
+
       console.log('Building ReVanced, please be patient!');
 
       const { stdout, stderr } = await actualExec(
@@ -260,15 +274,21 @@ async function checkForJavaADB () {
         await actualExec('adb install revanced/revanced.apk');
       }
 
-      console.log('Installing MicroG...');
+      if (adbExists) {
+        console.log('Installing MicroG...');
 
-      await actualExec(`adb install ${jarNames.microG}`);
+        await actualExec(`adb install ${jarNames.microG}`);
+      } else {
+        console.log('You now can install ReVanced and MicroG by transferring revanced/revanced.apk and revaned/microg.apk!')
+      }
+
       break;
     }
 
     default: {
       let patches = '';
       let useManualAPK;
+      let ytVersion
       if (!fs.existsSync('./revanced')) {
         fs.mkdirSync('./revanced');
       }
@@ -295,7 +315,9 @@ async function checkForJavaADB () {
       ];
       await downloadFiles(filesToDownload);
 
-      await getADBDeviceID();
+      if (adbExists) {
+        await getADBDeviceID();
+      }
 
       const getPatches = await actualExec(
         `java -jar ${jarNames.cli} -b ${jarNames.patchesJar} -l`
@@ -323,6 +345,7 @@ async function checkForJavaADB () {
       ]);
       for (const patch of patchesChoosed.patches) {
         if (patch === 'microg-support') {
+          ytVersion = await getYTVersion();
           patches += ' --mount';
         }
         patches += ` -e ${patch}`;
@@ -342,7 +365,7 @@ async function checkForJavaADB () {
       }
 
       if (!useManualAPK) {
-        await downloadYTApk();
+        await downloadYTApk(ytVersion);
       }
 
       console.log('Building ReVanced, please be patient!');
@@ -362,9 +385,14 @@ async function checkForJavaADB () {
         await actualExec('adb install revanced/revanced.apk');
       }
 
-      console.log('Installing MicroG...');
+      if (adbExists) {
+        console.log('Installing MicroG...');
 
-      await actualExec(`adb install ${jarNames.microG}`);
+        await actualExec(`adb install ${jarNames.microG}`);
+      } else {
+        console.log('You now can install ReVanced and MicroG by transferring revanced/revanced.apk and revaned/microg.apk!')
+      }
+
       break;
     }
   }
